move StringExtensions from TeeSharp.Common project to existing TeeSharp.Core->StringExtensions
