+++ TeeSharp.Common/TeeSharp.Common.csproj
@@ -13,4 +13,8 @@
       <PackageReference Include="Newtonsoft.Json" Version="13.0.1-beta1" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="src\Extensions\" />
+    </ItemGroup>
+
 </Project>
