+++ TeeSharp.Core/src/Extensions/StringExtensions.cs
@@ -15,5 +15,38 @@ namespace TeeSharp.Core.Extensions
                 ? source
                 : source.Substring(0, maxLength);
         }
+
+        public static ReadOnlySpan<char> SkipToWhitespaces(this ReadOnlySpan<char> input)
+        {
+            int index;
+            for (index = 0; index < input.Length; index++)
+            {
+                if (input[index] == ' ' ||
+                    input[index] == '\t' ||
+                    input[index] == '\n')
+                {
+                    break;
+                }
+            }
+
+            return input[index..];
+        }
+
+        public static ReadOnlySpan<char> SkipWhitespaces(this ReadOnlySpan<char> input)
+        {
+            int index;
+            for (index = 0; index < input.Length; index++)
+            {
+                if (input[index] != ' ' &&
+                    input[index] != '\t' &&
+                    input[index] != '\n' &&
+                    input[index] != '\r')
+                {
+                    break;
+                }
+            }
+
+            return input[index..];
+        }
     }
 }
